WIP: begin testing the visitor
This involved constructing a frankly ridiculous declarative macro
which expands into a mini-state-machine designed to validate that
we get all the visits and all the values that we want, in the order
we expect.

This actually almost works; it generates syntatically valid code.
Unfortunately, it's not quite ready yet; the list of items in the
input in line 26 is currently reversed as an unavoidable consequence
of the way we're building the enumeration. We need to re-enumerate
in the opposite direction in order for the state machine to work.